refactor: separate getVisiblePage to utils
