+++ src/pages/Catalog.tsx
@@ -1,4 +1,6 @@
 import { useEffect, useState } from "react"
+import { getVisiblePages } from "@/lib/utils"
+
 // Shadcn UI Imports
 import {
   InputGroup,
@@ -26,52 +28,24 @@ import { useProducts } from "../hooks/queries/useProducts"
 import CategoryChip from "../components/catalog/CategoryChip"
 import CategoryChipSkeleton from "../components/skeletons/catalog/CategoryChipSkeleton"
 import ProductCard from "@/components/catalog/ProductCard"
+import ProductCardSkeleton from "@/components/skeletons/catalog/ProductCardSkeleton"
 
 // Type Imports
 import type { GetProductParams } from "@/types/product.type"
 
 const LIMIT = 8
 
-// Helper function to determine visible page numbers for pagination
-const getVisiblePages = (currentPage: number, totalPages: number): Array<number | "ellipsis"> => {
-  if (totalPages <= 5) {
-    return Array.from({ length: totalPages }, (_, index) => index + 1)
-  }
-
-  // Use a Set to ensure unique page numbers and include the first, last, current, and adjacent pages
-  // * e.g., for currentPage = 5 and totalPages = 10, this will include 1, 4, 5, 6, and 10
-  const pages = new Set<number>([1, totalPages, currentPage, currentPage - 1, currentPage + 1])
-
-  // Filter out invalid page numbers and sort them in ascending order
-  // * e.g., if currentPage = 1, it will include 1, 2, and totalPages, but not 0 or negative numbers
-  const filteredPages = [...pages]
-    .filter((pageNumber) => pageNumber >= 1 && pageNumber <= totalPages)
-    .sort((a, b) => a - b)
-
-  const visiblePages: Array<number | "ellipsis"> = []
-
-  // Iterate through the filtered page numbers and insert "ellipsis" where there are gaps
-  // * e.g., for filteredPages = [1, 4, 5, 6, 10], this will produce [1, "ellipsis", 4, 5, 6, "ellipsis", 10]
-  filteredPages.forEach((pageNumber, index) => {
-    const previousPage = filteredPages[index - 1]
-    if (previousPage && pageNumber - previousPage > 1) {
-      visiblePages.push("ellipsis")
-    }
-    visiblePages.push(pageNumber)
-  })
-
-  return visiblePages
-}
-
 const Catalog = () => {
+  const [searchQuery, setSearchQuery] = useState("")
   const [selectedCategory, setSelectedCategory] = useState("all")
   const [page, setPage] = useState(1)
   const [productParams, setProductParams] = useState<GetProductParams>({
     limit: LIMIT,
   })
-  const [searchQuery, setSearchQuery] = useState("")
+
   const { data: categories, isLoading: isCategoriesLoading, isError: isCategoriesError } = useCategories()
   const { data: products, isLoading: isProductsLoading, isError: isProductsError } = useProducts(productParams, selectedCategory)
+
   const totalProducts = products?.total || 0
   const totalPages = Math.max(1, Math.ceil(totalProducts / (productParams.limit || LIMIT)))
   const visiblePages = getVisiblePages(page, totalPages)
